DockerPart2/crudappwithVolume/requirements.txt
autopep8==1.7.0
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycodestyle==2.9.1
pyparsing==3.0.9
python-dotenv==0.21.0
SQLAlchemy==1.4.42
toml==0.10.2
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
